WIP: allow fixed load balancer IPs for all Redis pods
This allows a new `sentinel.service.loadBalancerIPMappings` field. This
is an array that contains a set of static IPs. An IP from this set will
be used as the static load balancer IP for a pod, and that will persist
whether or not that pod is the primary.

`sentinel.service.loadBalancerIPMappings` must contain at least as many
entries as the number of replicas.